Day64/requirements.txt
Bootstrap_Flask
Requests
WTForms
Flask_WTF
Werkzeug
Flask
flask_sqlalchemy
SQLAlchemy
python-dotenv